adding albums

+++ src/main/java/models/Album.java
@@ -4,7 +4,7 @@ public class Album {
     private String title;
     private String artist;
     private int songCount;
-    private int length;
+    private float length;
     private String imgUrl;
 
     public Album(String title, String artist, int songCount, int length, String imgUrl) {
@@ -40,7 +40,7 @@ public class Album {
     }
 
     public int getLength() {
-        return length;
+        return (int) length;
     }
 
     public void setLength(int length) {
